+++ ARCaDIUS_Pump_Demo/ARCaDIUS_Serial.h
@@ -144,7 +144,7 @@ class ASerial {
     bool getValveState();
 
     int getMixer();
-    float getMixerSpeed();
+    int getMixerSpeed();
     bool getMixerDir();
 
     int getShutter();
